CleanUp and Added complex Statement as POC (GotoIfProgress)
