--- Playbook/Mutator.cs
@@ -1,6 +1,0 @@
-namespace Digdir.BDB.Dialogporten.ServiceProvider.Playbook;
-
-public class Mutator
-{
-    
-}
