+++ Playbook/PlaybookState.cs
@@ -1,5 +1,3 @@
-using System.Buffers.Text;
-using System.Runtime.CompilerServices;
 using System.Text;
 using System.Text.Json;
 using System.Text.Json.Nodes;
@@ -10,18 +8,19 @@ namespace Digdir.BDB.Dialogporten.ServiceProvider.Playbook;
 
 public class PlaybookState(Guid dialogId, int cursor, JsonArray patches)
 {
-
     public Guid DialogId { get; set; } = dialogId;
+
     public JsonArray Patches { get; set; } = patches;
+
     public int Cursor { get; set; } = cursor;
-    private static readonly JsonSerializerOptions _jsonSerializerOptions = new JsonSerializerOptions { WriteIndented = false };
+
+    private static readonly JsonSerializerOptions JsonSerializerOptions = new() { WriteIndented = false };
 
     public List<JsonPatchOperations_Operation>? CurrentPatch()
     {
         var jsonPatch = Patches[Cursor];
         return jsonPatch.Deserialize<List<JsonPatchOperations_Operation>>();
     }
-
 
     public static async Task<PlaybookState?> DecodeFromBase64(string base64)
     {
@@ -31,16 +30,14 @@ public class PlaybookState(Guid dialogId, int cursor, JsonArray patches)
         return JsonSerializer.Deserialize<PlaybookState>(Encoding.UTF8.GetString(decompressed));
     }
 
-
     public static Task<string> EncodeToBase64(PlaybookState playbookState)
     {
         return playbookState.EncodeToBase64();
     }
-    
-    
+
     public async Task<string> EncodeToBase64()
     {
-        var json = JsonSerializer.Serialize(this, _jsonSerializerOptions);
+        var json = JsonSerializer.Serialize(this, JsonSerializerOptions);
         var compressed = await CompressionExtensions.CompressBytesAsync(Encoding.UTF8.GetBytes(json));
 
         return Base64UrlEncoder.Encode(compressed);
